requirements.txt
streamlit
requests
load_dotenv